* unnecessary casts and semicolons cleanup
git-svn-id: https://j-wings.svn.sourceforge.net/svnroot/j-wings/trunk/wings2@2669 c8af5dc8-8010-0410-93ac-e0f74e17e1d8

+++ src/java/org/wings/script/JavaScriptListener.java
@@ -124,10 +124,10 @@ public class JavaScriptListener
             char c = code.charAt(i);
             endPos = i;
             if (lastChar == '{' && Character.isDigit(c)) {
-                int varIndex = (int) (c - '0');
+                int varIndex = (c - '0');
                 while (Character.isDigit(code.charAt(++i))) {
                     varIndex *= 10;
-                    varIndex += (int) (code.charAt(i) - '0');
+                    varIndex += (code.charAt(i) - '0');
                 }
                 c = code.charAt(i);
                 if (c == '}') {
